# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This server implements a simple API for a PARTNERS test.",
        "title": "RE PARTNERS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/packs",
    "paths": {
        "/config": {
            "put": {
                "description": "update the config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update the config",
                "operationId": "update-config",
                "parameters": [
                    {
                        "description": "Config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/{items}": {
            "get": {
                "description": "retrieve a number of packs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieve a number of packs",
                "operationId": "get-items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Items",
                        "name": "items",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Config": {
            "type": "object",
            "properties": {
                "packs_config": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}